+++ projects/hipcub/test/hipcub/test_hipcub_block_radix_sort.cpp
@@ -60,7 +60,7 @@ public:
 
 using Params = ::testing::Types<
 // Power of 2 BlockSize
-#if HIPCUB_IS_INT128_ENABLED
+#if _CCCL_HAS_INT128()
     params<__int128_t, __int128_t, 64U, 1>,
     params<__uint128_t, __uint128_t, 64U, 1>,
 #endif