+++ projects/hipcub/CHANGELOG.md
@@ -22,6 +22,7 @@ Full documentation for hipCUB is available at [https://rocm.docs.amd.com/project
 
 * Removed deprecated `hipcub:max` and `hipcub:min`. Use `hip::std::max` and `hip::std::min` instead.
 * Deprecated `hipcub::Swap`, use `rocprim::swap` instead.
+* Deprecated `HIPCUB_IS_INT128_ENABLED`, use `_CCCL_HAS_INT128()` instead.
 
 ### Changed